% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_addins.R
\name{Back2ForwardSlash}
\alias{Back2ForwardSlash}
\title{Replace \code{\\} with \code{/}}
\usage{
Back2ForwardSlash()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to replace
all single backslashes (\code{\\}) with single forwasdslashes (\code{/})
in the selected text.
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".
}
\seealso{
About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Replace selected symbols' addins: \code{\link{Back2doubleBackSlash}}
}

