% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_addins.R
\name{correct_writage_md}
\alias{correct_writage_md}
\title{[!] Correnct md text genereted with Writage}
\usage{
correct_writage_md(TEXT)
}
\arguments{
\item{x}{A sting.}
}
\value{
A corrected string.
}
\description{
[!] Correnct md text genereted with Writage
}
\examples{

    pattern     <- c( "\\\\`", "\\\\]",   "\\\\[",    "\\\\_",  "\\\\\\\\" )
    replacement <- c( "`",     "]",     "[",      "_",  "\\\\" )
}

