% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclose_addins.R
\name{enclose_with_asterisk}
\alias{enclose_with_asterisk}
\title{Enclose with \code{*}}
\usage{
enclose_with_asterisk()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to enclose the text with asterisk (\code{*}).
}
\seealso{
Other 'Enclose selected symbols' addins: \code{\link{enclose_with_asterisk2}},
  \code{\link{enclose_with_backtick}},
  \code{\link{enclose_with_italic_bold}},
  \code{\link{enclose_with_tilde2}},
  \code{\link{enclose_with_tilde}},
  \code{\link{enclose_with_underscore2}},
  \code{\link{enclose_with_underscore}},
  \code{\link{rs_equation_inline}}
}
