% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclose_addins.R
\name{enclose_with_asterisk2}
\alias{enclose_with_asterisk2}
\title{Enclose with \code{**}}
\usage{
enclose_with_asterisk2()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to enclose the text with double asterisk sign (\code{**}).
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".

}
\seealso{
About shortcut keys:
 \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Enclose selected symbols' addins: \code{\link{enclose_with_asterisk}},
  \code{\link{enclose_with_backtick}},
  \code{\link{enclose_with_dollar2}},
  \code{\link{enclose_with_dollar}},
  \code{\link{enclose_with_tilde2}},
  \code{\link{enclose_with_tilde}},
  \code{\link{enclose_with_underscore2}},
  \code{\link{enclose_with_underscore}}
}
