% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclose_addins.R
\name{enclose_with_backtick}
\alias{enclose_with_backtick}
\title{Enclose with \code{`}}
\usage{
enclose_with_backtick()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to enclose the text with back ticks sign (\code{`}).
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".

}
\seealso{
About shortcut keys:
 \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Enclose selected symbols' addins: \code{\link{enclose_with_asterisk2}},
  \code{\link{enclose_with_asterisk}},
  \code{\link{enclose_with_dolar2}},
  \code{\link{enclose_with_dolar}},
  \code{\link{enclose_with_underscore2}},
  \code{\link{enclose_with_underscore}}
}

