% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_formatting.R
\name{format_rmd}
\alias{format_rmd}
\alias{rmd_bold}
\alias{rmd_italic}
\alias{rmd_bold_ilalic}
\alias{rmd_bold2}
\alias{rmd_italic2}
\alias{rmd_code_inline}
\alias{rmd_r_code_inline}
\alias{rmd_superscript}
\alias{rmd_subscript}
\alias{rmd_strikethrough}
\title{R Markdown formatting}
\usage{
rmd_bold()

rmd_italic()

rmd_bold_ilalic()

rmd_bold2()

rmd_italic2()

rmd_code_inline()

rmd_r_code_inline()

rmd_superscript()

rmd_subscript()

rmd_strikethrough()
}
\description{
RStudio addins, which enclose selected text with symbols that give special interpretation
in R Markdown: \itemize{
  \item \code{rmd_bold()}:           between \code{**} and \code{**} as bold;
  \item \code{rmd_italic()}:         between \code{_} and \code{_} as italic;
  \item \code{rmd_bold_italic()}:    between \code{**_} and \code{_**}as bold italic;
  \item \code{rmd_bold2()}:          between \code{__} and \code{__} as bold;
  \item \code{rmd_italic2()}:        between \code{*} and \code{*} as italic;
  \item \code{rmd_code_inline()}:    between backticks (``) as inline code;
  \item \code{rmd_r_code_inline()}:  formats text as inline R code to evaluate;
  \item \code{rmd_superscript()}:    between \code{^} and \code{^} as superscript;
  \item \code{rmd_subscript()}:      between \code{~} and \code{~} as subscript;
  \item \code{rmd_strikethrough()}:  between \code{~~} and \code{~~} as strike-through text.
}
}
\seealso{
Other R Markdown formatting addins: \code{\link{rmd_code_block}},
  \code{\link{rmd_equations}}, \code{\link{rmd_list}}
}
