% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_arrows.R
\name{insert_arrows}
\alias{insert_arrows}
\alias{rs_insert_arrow_rl}
\title{Insert arrow (->, ->>, <-, <<-)}
\usage{
rs_insert_arrow_rl()
}
\description{
RStudio addin to insert either \code{->}, \code{->>},
\code{<-} or \code{<<-} at the cursor position.
}
\note{
use hotkeys combination \code{ALT} + \code{-} to insert \code{<-} in RStudio.
}
\seealso{
Assignment operators \link[base]{assignOps}.\cr

\dontrun{
\donttest{
library(spAddins)

rs_insert_arrow_rl()
## <-

rs_insert_arrow_rl2()
## <<-

rs_insert_arrow_lr()
## ->

rs_insert_arrow_lr2()
## ->>
}}

Other 'Insert at cursor position' addins: \code{\link{insert_magrittr_operators}},
  \code{\link{insert_symbol_sequences}},
  \code{\link{rs_insert_infix_R}},
  \code{\link{rs_insert_infix_in}},
  \code{\link{rs_insert_infix_not_in}},
  \code{\link{rs_insert_matrix_multiplication}}
}
