% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_Line_addins.R
\name{insert_chunk_python_Addin}
\alias{insert_chunk_python_Addin}
\title{Insert python's code chunk for conspects}
\usage{
insert_chunk_python_Addin()
}
\description{
Call this function as an addin to insert python's code chunk for conspects at the cursor position.
}
\seealso{
Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_chunk_r_Addin}},
  \code{\link{insert_ds_line_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{insert_ss_line_Addin}},
  \code{\link{insert_sw_line_Addin}}
}

