% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_magrittr.R
\name{insert_magrittr_operators}
\alias{insert_magrittr_operators}
\alias{rs_insert_pipe}
\alias{rs_insert_update_pipe}
\alias{rs_insert_exposition_pipe}
\alias{rs_insert_tee_pipe}
\title{Insert \%>\%, \%<>\%, \%$\% and \%T>\%}
\usage{
rs_insert_pipe()

rs_insert_update_pipe()

rs_insert_exposition_pipe()

rs_insert_tee_pipe()
}
\description{
Call this function as an addin to insert \code{ \%>\% } at the cursor position.
}
\note{
Operator does not work unless package \pkg{magrittr} or
equivalent is loaded.
}
\seealso{
Pipeline operator \link[magrittr]{\%>\%}.\cr
Compound assignment pipe operator \link[magrittr]{\%<>\%}.\cr
Exposition pipe operator \link[magrittr]{\%$\%} and operator \link[base]{$}.\cr
Tee operator \link[magrittr]{\%T>\%}.\cr

Other 'Insert at cursor position' addins: \code{\link{insert_arrows}},
  \code{\link{insert_symbol_sequences}},
  \code{\link{rs_insert_infix_R}},
  \code{\link{rs_insert_infix_in}},
  \code{\link{rs_insert_infix_not_in}},
  \code{\link{rs_insert_matrix_multiplication}}
}
