% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_Line_addins.R
\name{insert_sw_line_Addin}
\alias{insert_sw_line_Addin}
\title{Insert single wavy (SW) line}
\usage{
insert_sw_line_Addin()
}
\description{
Call this function as an addin to insert \code{# ~~~~~} at the cursor position.
}
\seealso{
Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_ds_line_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{insert_ss_line_Addin}}
}

