% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arrows_insert_addins.R
\name{insertArrowLR2_Addin}
\alias{insertArrowLR2_Addin}
\alias{insertArrowRL_Addin}
\alias{insertArrowRL2_Addin}
\title{Insert arrow (->, ->>, <<-)}
\usage{
insertArrowLR2_Addin()

insertArrowRL_Addin()

insertArrowRL2_Addin()
}
\description{
Call these functions as addins to insert either \code{->},\code{->>},
or \code{<<-} at the cursor position.
}
\note{
use hotkeys combination \code{ALT + -} to insert \code{<-} in RStudio.
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".

\dontrun{
\donttest{
library(spAddins)

insertArrowLR2_Addin()
## <<-

insertArrowRL_Addin()
## ->

insertArrowRL2_Addin()
## ->>
}}

}
\seealso{
Assignment operators \link[base]{assignOps}.\cr
 About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Insert at cursor position' addins: \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPasteRebus}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_ds_line_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{insert_ss_line_Addin}},
  \code{\link{insert_sw_line_Addin}}
}
