% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_addins.R
\name{insertIfNULL_Addin}
\alias{insertIfNULL_Addin}
\title{Insert \%if.NULL\%}
\usage{
insertIfNULL_Addin()
}
\description{
Call this function as an addin to insert \code{ \%if.NULL\% } at the cursor position.
}
\note{
Operator does not work unless library which contains this operator is loaded.
}
\examples{

# To call the functions with keyboard shortcuts explore link "spAddins"
# in section "See also".
}
\seealso{
About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}}
}

