% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_addins.R
\name{insertIn_Addin}
\alias{insertIn_Addin}
\title{Insert \%in\%}
\usage{
insertIn_Addin()
}
\description{
Call this function as an addin to insert \code{ \%in\% } at the cursor position.
}
\seealso{
Operator \link[base]{\%in\%}

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}}
}

