% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_addins.R
\name{insertMatMuliplication_Addin}
\alias{insertMatMuliplication_Addin}
\title{Insert \%*\%}
\usage{
insertMatMuliplication_Addin()
}
\description{
Call this function as an addin to insert \code{ \%*\% } at the cursor position.
}
\seealso{
Operator \link[base]{\%*\%} for matrix multiplication.

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}}
}

