% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_addins.R
\name{insertPaste0_Addin}
\alias{insertPaste0_Addin}
\title{Insert \%++\%}
\usage{
insertPaste0_Addin()
}
\description{
Call this function as an addin to insert \code{ \%++\% } at the cursor position.
}
\note{
Operator does not work unless library which contains this operator is loaded.
}
\examples{
# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".

}
\seealso{
About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.\cr

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_ds_line_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{insert_ss_line_Addin}},
  \code{\link{insert_sw_line_Addin}}
}

