% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_addins.R
\name{insertPasteRebus}
\alias{insertPasteRebus}
\title{Insert \%R\%}
\usage{
insertPasteRebus()
}
\description{
Call this function as an addin to insert \code{ \%R\% } at the cursor position.
}
\note{
Operator does not work unless library which contains this operator is loaded.
}
\examples{
# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".

}
\seealso{
About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}. \cr

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{rs_insert_ds_line}},
  \code{\link{rs_insert_ss_line}},
  \code{\link{rs_insert_sw_line}}
}
