% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_operators_insert_addins.R
\name{insertTeeOperator_Addin}
\alias{insertTeeOperator_Addin}
\title{Insert \%T>\%}
\usage{
insertTeeOperator_Addin()
}
\description{
Call this function as an addin to insert \code{ \%T>\% } at the cursor position.
}
\note{
Operator does not work unless library \pkg{magrittr} or
equivalent is loaded.
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".
}
\seealso{
Tee operator \link[magrittr]{\%T>\%}\cr
  About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}}
}

