% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_symbol.R
\name{repeat_symbol}
\alias{repeat_symbol}
\title{Repeat sequence of symbols}
\usage{
repeat_symbol(text, length.out)
}
\arguments{
\item{text}{(character) The symbol (or sequence of symbols) to be repeated
until desired length of string.}

\item{length.out}{(integer) The length of the sequence.}
}
\value{
Sting of defined length.
}
\description{
Repeat sequence of symbols
}
\examples{

repeat_symbol(".", 10)

repeat_symbol("..+", 10)
}
