% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_slash.R
\name{replace_slash}
\alias{replace_slash}
\alias{rs_replace_slash_bs2d}
\alias{rs_replace_slash_bd2s}
\alias{rs_replace_slash_b2fw}
\alias{rs_replace_slash_fw2b}
\title{Replace slash.}
\usage{
rs_replace_slash_bs2d()

rs_replace_slash_bd2s()

rs_replace_slash_b2fw()

rs_replace_slash_fw2b()
}
\description{
RStudio add-in to manage various types of slashes.
Select a piece of text with a cursor and do the necessary replacement
operation in the selected text: \itemize{
   \item \code{rs_replace_slash_bs2d} single back-slash into double back-slash;
   \item \code{rs_replace_slash_bd2s} double back-slash into single back-slash;
   \item \code{rs_replace_slash_b2fw} back-slash into forward-slash;
   \item \code{rs_replace_slash_fw2b} forward-slash into back-slash.
}
}
