% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_code_block.R
\name{rmd_code_block}
\alias{rmd_code_block}
\alias{rmd_code_block_r}
\alias{rmd_code_block}
\title{Convert rows into the block of code}
\usage{
rmd_code_block_r()

rmd_code_block()
}
\description{
RStudio add-in to insert selected lines into code block: \itemize{
   \item \code{rmd_r_code_block()} - R code block;
   \item \code{rmd_code_block()} - verbatim code block.
}
\code{rs_enclose_all_with_lines} - function that adds lines above and below the selection.
}
