% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_equations.R
\name{rmd_equations}
\alias{rmd_equations}
\alias{rmd_equation_inline}
\alias{rmd_equation_block}
\alias{rmd_equation_block2}
\title{As LaTeX equation}
\usage{
rmd_equation_inline()

rmd_equation_block()

rmd_equation_block2()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to enclose the text to interpret as inline equation (\code{$.$}) or
block equotion  (e.g., \code{$$.$$} or \code{\[.\]}) in R Markdown.
}
\seealso{
Other R Markdown formatting addins: \code{\link{format_rmd}},
  \code{\link{rmd_code_block}}, \code{\link{rmd_list}}
}
