% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lists.R
\name{rmd_list}
\alias{rmd_list}
\alias{rmd_list_unordered}
\alias{rmd_list_numbered}
\alias{rmd_list_lettered}
\alias{rmd_master_list}
\title{R Markdown lists}
\usage{
rmd_list(type = "unordered", level = 1)

rmd_list_unordered(variables)

rmd_list_numbered(variables)

rmd_list_lettered(variables)

rmd_master_list(variables)
}
\seealso{
Other R Markdown formatting addins: \code{\link{format_rmd}},
  \code{\link{rmd_code_block}}, \code{\link{rmd_equations}}
}
