% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclose_addins.R
\name{rs_equation_inline}
\alias{rs_equation_inline}
\alias{rs_equation_block}
\alias{rs_equation_block2}
\title{Enclose with \code{$}}
\usage{
rs_equation_inline()

rs_equation_block()

rs_equation_block2()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to enclose the text to interpret as inline equation (\code{$.$}) or
block equotion  (e.g., \code{$$.$$} or \code{\[. \]}) in R Markdown.
}
\seealso{
Other 'Enclose selected symbols' addins: \code{\link{enclose_with_asterisk2}},
  \code{\link{enclose_with_asterisk}},
  \code{\link{enclose_with_backtick}},
  \code{\link{enclose_with_italic_bold}},
  \code{\link{enclose_with_tilde2}},
  \code{\link{enclose_with_tilde}},
  \code{\link{enclose_with_underscore2}},
  \code{\link{enclose_with_underscore}}

Other 'Enclose selected symbols' addins: \code{\link{enclose_with_asterisk2}},
  \code{\link{enclose_with_asterisk}},
  \code{\link{enclose_with_backtick}},
  \code{\link{enclose_with_italic_bold}},
  \code{\link{enclose_with_tilde2}},
  \code{\link{enclose_with_tilde}},
  \code{\link{enclose_with_underscore2}},
  \code{\link{enclose_with_underscore}}

Other 'Enclose selected symbols' addins: \code{\link{enclose_with_asterisk2}},
  \code{\link{enclose_with_asterisk}},
  \code{\link{enclose_with_backtick}},
  \code{\link{enclose_with_italic_bold}},
  \code{\link{enclose_with_tilde2}},
  \code{\link{enclose_with_tilde}},
  \code{\link{enclose_with_underscore2}},
  \code{\link{enclose_with_underscore}}
}
