% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_ind_first_selected_col.R
\name{rs_get_ind}
\alias{rs_get_ind}
\alias{rs_get_ind_first_selected_col}
\alias{rs_get_ind_last_selected_col}
\alias{rs_get_ind_first_selected_row}
\alias{rs_get_ind_last_selected_row}
\title{Get indices from the selection.}
\usage{
rs_get_ind_first_selected_col(context = get_context())

rs_get_ind_last_selected_col(context = get_context())

rs_get_ind_first_selected_row(context = get_context())

rs_get_ind_last_selected_row(context = get_context())
}
\arguments{
\item{context}{Object with context of active R Studio document
(class \code{document_context}).}
}
\description{
Get index of the first/last row/column in the selection.
}
