% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_ind_first_selected_col.R
\name{rs_get_ind_first_selected_col}
\alias{rs_get_ind_first_selected_col}
\title{Get index of the first column in selection}
\usage{
rs_get_ind_first_selected_col()
}
\description{
Get index of the first column in selection
}
