% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_insert_at_row_start.R
\name{rs_insert_before_first_selected_row}
\alias{rs_insert_before_first_selected_row}
\title{Insert text at the beginning of the first selected row.}
\usage{
rs_insert_before_first_selected_row(text = "", ensure_blank_above = FALSE,
  context = get_context())
}
\arguments{
\item{text}{(character) The text to add.}

\item{ensure_blank_above}{(logical)
If \code{TRUE}, checks if there is a
blank line above/bolow the selection. If line is not blank,
 a blank line is added. If \code{FALSE}, blank line is not added.}

\item{context}{Object with context of active R Studio document
(class \code{document_context}).}
}
\description{
Insert text at the beginning of the first selected row.
}
