% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_Line_addins.R
\name{rs_insert_ds_line}
\alias{rs_insert_ds_line}
\title{Insert double straight (DS) line}
\usage{
rs_insert_ds_line(start = selection_start_column(), end = 70)
}
\arguments{
\item{start}{(numeric) The number of column the line begins at.}

\item{end}{(numeric) The number of column the line ends at.}
}
\description{
Call this function as an addin to insert \code{======} at the cursor position.
}
\seealso{
Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertExPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPasteRebus}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{rs_insert_ss_line}},
  \code{\link{rs_insert_sw_line}}
}
