% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_infix_in}
\alias{rs_insert_infix_in}
\title{Insert \%in\%.}
\usage{
rs_insert_infix_in()
}
\description{
Call this function as an add-in to insert \code{ \%in\% } at the cursor position.
}
\seealso{
Operator \link[base]{\%in\%}.

Other 'Insert at cursor position' add-ins: \code{\link{insert_magrittr_operators}},
  \code{\link{rs_insert_infix_R}},
  \code{\link{rs_insert_infix_not_in}},
  \code{\link{rs_insert_matrix_multiplication}}
}
