% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_infix_R}
\alias{rs_insert_infix_R}
\title{Insert \%R\%}
\usage{
rs_insert_infix_R()
}
\description{
Call this function as an addin to insert \code{ \%R\% } at the cursor position.
}
\note{
Operator does not work unless library which contains this operator is loaded.
}
\seealso{
Operator \link[rebus.base]{\%R\%} in \pkg{rebus}.

Other 'Insert at cursor position' addins: \code{\link{insert_arrows}},
  \code{\link{insert_magrittr_operators}},
  \code{\link{insert_symbol_sequences}},
  \code{\link{rs_insert_infix_in}},
  \code{\link{rs_insert_infix_not_in}},
  \code{\link{rs_insert_matrix_multiplication}}
}
