% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_matrix_multiplication}
\alias{rs_insert_matrix_multiplication}
\title{Insert \%*\%}
\usage{
rs_insert_matrix_multiplication()
}
\description{
Call this function as an add-in to insert \code{ \%*\% } at the cursor position.
}
\seealso{
Operator \link[base]{\%*\%} for matrix multiplication.

Other 'Insert at cursor position' add-ins: \code{\link{insert_magrittr_operators}},
  \code{\link{rs_insert_infix_R}},
  \code{\link{rs_insert_infix_in}},
  \code{\link{rs_insert_infix_not_in}}
}
