% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_addins.R
\name{rs_replace_backslash_with_forwasdslash}
\alias{rs_replace_backslash_with_forwasdslash}
\title{Replace \code{\\} with \code{/}}
\usage{
rs_replace_backslash_with_forwasdslash()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to replace
all single back-slashes (\code{\\}) with single forwasd-slashes (\code{/})
in the selected text.
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".
}
\seealso{
About shortcut keys:
\href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Replace selected symbols' addins: \code{\link{rs_replace_double_with_single_backslash}},
  \code{\link{rs_replace_single_with_double_backslash}}
}
