% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_replace_in_selection.R
\name{rs_replace_in_selection}
\alias{rs_replace_in_selection}
\title{Replace text in selection}
\usage{
rs_replace_in_selection(pattern, replacement)
}
\arguments{
\item{pattern}{(character) A fixted pattern of text to be replaced
(not a regular expression).}

\item{replacement}{(character) The replacement text.}
}
\description{
Replace text in selection
}
