% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_addins.R
\name{rs_replace_single_with_double_backslash}
\alias{rs_replace_single_with_double_backslash}
\title{Replace \code{\\} with \code{\\\\}}
\usage{
rs_replace_single_with_double_backslash()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to replace all single backslashes (\code{\\}) with double backslashes
(\code{\\\\}) in the seleceted text.
}
\seealso{
Other 'Replace selected symbols' addins: \code{\link{rs_replace_backslash_with_forwasdslash}},
  \code{\link{rs_replace_double_with_single_backslash}}
}
