% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_r_code.R
\name{rs_tidy_r_code}
\alias{rs_tidy_r_code}
\title{Addin to tidy up several aspects of an R code in a selection}
\usage{
rs_tidy_r_code()
}
\description{
Addin to tidy up several aspects of an R code in a selection
}
