% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_spAddins-package.R
\docType{package}
\name{spAddins}
\alias{spAddins}
\alias{spAddins-package}
\title{A Set of RStudio Addins}
\description{
\code{spAddins} is an R package, that provides a set of RStudio
 addins. Its recommended to create
 \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}
 and call the functions from this package using those shortcuts.\cr

Learn more about \href{https://rstudio.github.io/rstudioaddins/}{RStudio addins}
and especially how to use them in combination with
\href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.\cr


\code{URL:} \url{https://github.com/GegznaV/spAddins} \cr
\code{Bug reports and suggestions:} \url{https://github.com/GegznaV/spAddins/issues} \cr
\code{License:} MIT \cr
\code{Author:} Vilmantas Gegzna  \cr
}

