% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_r_code.R
\name{tidy_r_code_Addin}
\alias{tidy_r_code_Addin}
\title{Tidy up several aspects of an R code in a selection}
\usage{
tidy_r_code_Addin()
}
\description{
Tidy up several aspects of an R code in a selection
}
