% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_addins.R
\name{Back2doubleBackSlash}
\alias{Back2doubleBackSlash}
\title{Replace \code{\\} with \code{\\\\}}
\usage{
Back2doubleBackSlash()
}
\description{
Select a piece of text with a cursor and call this function as an addin
to replace all single backslashes (\code{\\}) with double backslashes
(\code{\\\\}) in the seleceted text.
}
\seealso{
Other 'Replace selected symbols' addins: \code{\link{Back2ForwardSlash}}
}

