% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_elements.R, R/insert_arrows.R
\name{insert_arrows}
\alias{insert_arrows}
\alias{html_insert_space}
\alias{rs_insert_arrow_rl}
\alias{rs_insert_arrow_rl2}
\alias{rs_insert_arrow_lr}
\alias{rs_insert_arrow_lr2}
\title{Insert HTML elements.}
\usage{
html_insert_space()

rs_insert_arrow_rl()

rs_insert_arrow_rl2()

rs_insert_arrow_lr()

rs_insert_arrow_lr2()
}
\description{
RStudio add-in to insert HTML elements

RStudio add-in to insert either \code{->}, \code{->>},
\code{<-} or \code{<<-} at the cursor position.
}
\note{
use hotkeys combination \code{ALT} + \code{-} to insert \code{<-} in RStudio.
}
\examples{
\dontrun{
\donttest{
library(spAddins)

rs_insert_arrow_rl()
## <-

rs_insert_arrow_rl2()
## <<-

rs_insert_arrow_lr()
## ->

rs_insert_arrow_lr2()
## ->>
}
}
}
\seealso{
Assignment operators \link[base]{assignOps}.\cr

Other 'Insert at cursor position' addins: \code{\link{insert_symbol_sequences}}

Other 'Insert at cursor position' addins: \code{\link{insert_symbol_sequences}}
}
\concept{'Insert at cursor position' addins}
