% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_magrittr.R
\name{insert_magrittr_operators}
\alias{insert_magrittr_operators}
\alias{rs_insert_pipe}
\alias{rs_insert_update_pipe}
\alias{rs_insert_exposition_pipe}
\alias{rs_insert_tee_pipe}
\title{Insert magrittr operators.}
\usage{
rs_insert_pipe()

rs_insert_update_pipe()

rs_insert_exposition_pipe()

rs_insert_tee_pipe()
}
\description{
These add-ins insert operators \code{\%>\%}, \code{\%<>\%}, \code{\%$\%},
and \code{\%T>\%} at the cursor position.
The operators are defined in package \pkg{magrittr}.
}
\note{
Operator does not work unless package \pkg{magrittr} or
equivalent is loaded.
}
\seealso{
\itemize{
   \item Forward pipeline operator \link[magrittr]{\%>\%}
   \item Compound assignment pipe operator \link[magrittr]{\%<>\%}
   \item Exposition pipe operator \link[magrittr]{\%$\%} and operator \link[base]{$}
   \item Tee operator \link[magrittr]{\%T>\%}
}

Other 'Insert at cursor position' add-ins: \code{\link{rs_insert_infix_R}},
  \code{\link{rs_insert_infix_in}},
  \code{\link{rs_insert_infix_not_in}},
  \code{\link{rs_insert_matrix_multiplication}}
}
\concept{'Insert at cursor position' add-ins}
