% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_operators_insert_addins.R
\name{insertExPipe_Addin}
\alias{insertExPipe_Addin}
\title{Insert \%$\%}
\usage{
insertExPipe_Addin()
}
\description{
Call this function as an addin to insert \code{ \%$\% } at the cursor position.
}
\note{
Operator does not work unless library \pkg{magrittr} or
equivalent is loaded.
}
\examples{

# To call the functions with keyboard shortcuts explore
# link "keyboard shortcuts" in section "See also".
}
\seealso{
Exposition pipe operator \link[magrittr]{\%$\%},
 operator \link[base]{$}.\cr
  About shortcut keys: \href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{keyboard shortcuts}.

Other 'Insert at cursor position' addins: \code{\link{insertArrowLR2_Addin}},
  \code{\link{insertCompAssignPipe_Addin}},
  \code{\link{insertIfNULL_Addin}},
  \code{\link{insertIn_Addin}},
  \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertNotIn_Addin}},
  \code{\link{insertPaste0_Addin}},
  \code{\link{insertPasteRebus}},
  \code{\link{insertPaste_Addin}},
  \code{\link{insertPipeline_Addin}},
  \code{\link{insertTeeOperator_Addin}},
  \code{\link{insert_if_null_Addin}},
  \code{\link{insert_if_null_or_len0_Addin}},
  \code{\link{rs_insert_ds_line}},
  \code{\link{rs_insert_ss_line}},
  \code{\link{rs_insert_sw_line}}
}
