% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_equations.R
\name{rmd_equations}
\alias{rmd_equations}
\alias{rmd_equation_inline}
\alias{rmd_equation_block}
\alias{rmd_equation_block2}
\title{Format as LaTeX equation.}
\usage{
rmd_equation_inline()

rmd_equation_block()

rmd_equation_block2()
}
\description{
Select a piece of text with a cursor and call these functions as an add-in
to enclose the text with special symbols. Text between \code{$.$} is
interpreted as an inline equation and between \code{$$.$$} or \code{\[.\]}
as a block equation in R Markdown.
}
\seealso{
Other R Markdown formatting add-ins: \code{\link{format_rmd}},
  \code{\link{rmd_code_block}}, \code{\link{rmd_list}}
}
