% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_lists.R
\name{rmd_list}
\alias{rmd_list}
\alias{rmd_list_unordered}
\alias{rmd_list_unordered_2}
\alias{rmd_list_numbered}
\alias{rmd_list_numbered_2}
\alias{rmd_list_lettered}
\alias{rmd_list_lettered_2}
\alias{rmd_master_list}
\title{Format text as R Markdown list}
\usage{
rmd_list(type = "unordered", level = 1)

rmd_list_unordered()

rmd_list_unordered_2()

rmd_list_numbered()

rmd_list_numbered_2()

rmd_list_lettered()

rmd_list_lettered_2()

rmd_master_list()
}
\arguments{
\item{type}{(character) the type of list "unordered", "numbered", "lettered",  "LETTERED", "master".}

\item{level}{(integer) the level of list.}
}
\description{
RStudio add-ins which convert text into R Markdown lists.
For the first-level lists: \itemize{
  \item \code{rmd_list()} - the main function, that make lists;
  \item \code{rmd_unordered_list()} - unordered list;
  \item \code{rmd_numbered_list()} - numbered list;
  \item \code{rmd_lettered_list()} - lettered list (non-capital English letters);
  \item \code{rmd_master_list()} - master list (which numbering continues throughout the document).
  }
}
\seealso{
Other R Markdown formatting add-ins: \code{\link{format_rmd}},
  \code{\link{rmd_equations}}
}
