% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_tables.R
\name{rmd_tables}
\alias{rmd_tables}
\alias{rmd_table_1}
\alias{rmd_table_2}
\alias{rmd_table_3}
\alias{rmd_table_4}
\alias{rmd_table_5}
\title{Insert R Markdown tables}
\usage{
rmd_table_1()

rmd_table_2()

rmd_table_3()

rmd_table_4()

rmd_table_5()
}
\description{
Insert examples of R Markdown tables.
}
