% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_enclose_with_lines.R
\name{rs_enclose_all_with_lines}
\alias{rs_enclose_all_with_lines}
\alias{rs_enclose_first_row_with_lines}
\title{Enclose selection with lines.}
\usage{
rs_enclose_all_with_lines(above = NA, below = NA,
  ensure_blank_above = FALSE, context = get_context())

rs_enclose_first_row_with_lines(above = NULL, below = NULL,
  ensure_blank_above = FALSE, context = get_context())
}
\arguments{
\item{above}{(character) Text to be inserted in the row above the selection.}

\item{below}{(character) Text to be inserted in the row below the selection
(or the first row).}

\item{ensure_blank_above}{(logical)
If \code{TRUE}, checks if there is a
blank line above the selection. If line is not blank, a blank line
is added. If \code{FALSE}, blank line is not added.}

\item{context}{Object with context of active R Studio document
(class \code{document_context}).}
}
\description{
Enclose selected rows with lines above and below: \itemize{
    \item \code{rs_enclose_all_with_lines()} - all selected rows;
    \item \code{rs_enclose_first_row_with_lines()} - the first selected row only.
}
}
