% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_enclose_selection_with.R
\name{rs_enclose_selection_with}
\alias{rs_enclose_selection_with}
\title{Enclose selection of text with indicated symbols}
\usage{
rs_enclose_selection_with(symbol = "", symbol_before = symbol,
  symbol_after = symbol)
}
\arguments{
\item{symbol}{(character) A sequence of symbols to add on both sides of selection.}

\item{symbol_before}{(character) A sequence of symbols to before the selection
(overrides value of \code{symbol}).}

\item{symbol_after}{(character) A sequence of symbols to add after the selection
(overrides value of \code{symbol}).}
}
\description{
Enclose selection of text with indicated symbols
}
