% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_r_code.R
\name{tidy_r_code}
\alias{tidy_r_code}
\title{Tidy up several aspects of an R code in a selection}
\usage{
tidy_r_code(str)
}
\arguments{
\item{str}{a sting}
}
\description{
Other aspects are corrected by clicking \strong{Ctrl + Shift + A} and
\strong{Ctrl + I} in RStudio.
}
