% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_r_code.R
\name{tidy_r_code}
\alias{tidy_r_code}
\title{Tidy up several aspects of an R code}
\usage{
tidy_r_code(str)
}
\arguments{
\item{str}{a sting}
}
\description{
Tidy up several aspects of an R code in a selection: #' \enumerate{
   \item Remove a space before a bracket;
   \item Make space betweeb `for`/ `if` and opening of its condition;
   \item Remove speces before and after `:`, `::`, and `:::`.
}
Other aspects are corrected by clicking \strong{Ctrl + Shift + A} and
\strong{Ctrl + I} in RStudio.
}
